+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/roomlist/room/RoomNameAndTime.kt
@@ -3,10 +3,8 @@ package de.connect2x.messenger.compose.view.roomlist.room
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -17,16 +15,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import de.connect2x.messenger.compose.view.DI
 import de.connect2x.messenger.compose.view.Tooltip
 import de.connect2x.messenger.compose.view.common.PlaceholderHighlight
 import de.connect2x.messenger.compose.view.common.TooltipText
 import de.connect2x.messenger.compose.view.common.fade
 import de.connect2x.messenger.compose.view.common.icons.UnencryptedIcon
 import de.connect2x.messenger.compose.view.common.placeholder
-import de.connect2x.messenger.compose.view.get
-import de.connect2x.messenger.compose.view.i18n.I18nView
-import de.connect2x.messenger.compose.view.util.TextLabel
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListElementViewModel
 
 @Composable
@@ -66,8 +60,6 @@ fun RowScope.RoomTime(roomListElementViewModel: RoomListElementViewModel, modifi
 @Composable
 fun RoomNameAndTime(roomListElementViewModel: RoomListElementViewModel) {
     val roomName = roomListElementViewModel.roomName.collectAsState().value
-    val hasLeft = roomListElementViewModel.hasLeft.collectAsState().value
-    val i18n = DI.get<I18nView>()
 
     Row(
         modifier = Modifier.placeholder(
@@ -80,10 +72,6 @@ fun RoomNameAndTime(roomListElementViewModel: RoomListElementViewModel) {
     ) {
         Row(Modifier.fillMaxWidth().weight(1.0f, false), verticalAlignment = Alignment.CenterVertically) {
             RoomName(roomName = roomName)
-            Spacer(Modifier.size(7.dp))
-            if (hasLeft == true) {
-                TextLabel(i18n.commonArchived())
-            }
         }
 
         RoomTime(roomListElementViewModel)