+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/settings/RoomSettingsViewModelTest.kt
@@ -165,32 +165,6 @@ class RoomSettingsViewModelTest : ShouldSpec() {
             every { userServiceMock.canSetPowerLevelToMax(eq(roomId), any()) } returns MutableStateFlow(100)
         }
 
-        should("go back to the room list view when leaving the room successfully") {
-            every {
-                userServiceMock.getAccountData(eq(PushRulesEventContent::class), any())
-            } returns MutableStateFlow(null)
-            everySuspend {
-                roomsApiClientMock.leaveRoom(
-                    eq(roomId),
-                    any(),
-                    eqNull()
-                )
-            } returns
-                    Result.success(Unit)
-            val onBackMock = mock<Function0<Unit>>()
-            val cut = roomSettingsViewModel(coroutineContext, onBackMock)
-
-            cut.leaveRoom()
-            testCoroutineScheduler.advanceUntilIdle()
-
-            verifySuspend {
-                roomsApiClientMock.leaveRoom(eq(roomId), any(), eqNull())
-                onBackMock()
-            }
-
-            cancelNeverEndingCoroutines()
-        }
-
         should("show an error message when trying to leave a room and we are not connected") {
             every {
                 userServiceMock.getAccountData(eq(PushRulesEventContent::class), any())